requirements-dev.txt
Django==2.2
django-bootstrap4==0.0.8
django-braces==1.13.0
django-cors-headers==2.5.3
django-flatpickr==1.0.1
djangorestframework==3.9.2
mysql-connector==2.2.9
mysqlclient==2.1.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.3
